Reorganize and improve documentation.
